+++ src/components/layouts/BasicLayout/index.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useRef } from 'react';
 import { Outlet } from 'react-router-dom';
 
-import FABContainer from '@/components/layouts/FAB';
+import FABContainer from '@/components/layouts/FAB/FABContainer';
 import Header from '@/components/layouts/Header';
 import TabBar from '@/components/layouts/TabBar';
