+++ src/components/layouts/FAB/index.tsx
@@ -1,56 +1,13 @@
 import styled from '@emotion/styled';
-import { useEffect, useState } from 'react';
 
 import AddIcon from '@/assets/icons/add.svg?react';
 import ArrowUpwardIcon from '@/assets/icons/arrow-upward.svg?react';
-import type { Mode } from '@/types';
-import { TABBAR_HEIGHT } from '../TabBar';
-
-interface FABContainerProps {
-  mode: Mode;
-  scrollContainerRef: React.RefObject<HTMLElement>;
-}
-
-const FABContainer = ({ mode, scrollContainerRef }: FABContainerProps) => {
-  const [showScrollToTopButton, setShowScrollToTopButton] = useState(false);
-
-  useEffect(() => {
-    const currentContainer = scrollContainerRef.current;
-
-    const handleScroll = () => {
-      if (currentContainer?.scrollTop && currentContainer.scrollTop > 0) {
-        setShowScrollToTopButton(true);
-      } else {
-        setShowScrollToTopButton(false);
-      }
-    };
-
-    if (currentContainer) {
-      currentContainer.addEventListener('scroll', handleScroll);
-    }
-
-    return () => {
-      if (currentContainer) {
-        currentContainer.removeEventListener('scroll', handleScroll); // 언마운트
-      }
-    };
-  }, [scrollContainerRef]);
-
-  return (
-    <Wrapper>
-      {showScrollToTopButton && <ScrollToTopButton scrollContainerRef={scrollContainerRef} />}
-      {mode === 'seller' && <PostButton />}
-    </Wrapper>
-  );
-};
-
-export default FABContainer;
 
 interface ScrollToTopButtonProps {
   scrollContainerRef: React.RefObject<HTMLElement>;
 }
 
-const ScrollToTopButton = ({ scrollContainerRef }: ScrollToTopButtonProps) => {
+export const ScrollToTopButton = ({ scrollContainerRef }: ScrollToTopButtonProps) => {
   const scrollToTop = () => {
     if (scrollContainerRef.current) {
       scrollContainerRef.current.scrollTo({ top: 0, behavior: 'smooth' });
@@ -64,23 +21,13 @@ const ScrollToTopButton = ({ scrollContainerRef }: ScrollToTopButtonProps) => {
   );
 };
 
-const PostButton = () => {
+export const PostButton = () => {
   return (
     <StyledPostButton>
       <AddIcon />
     </StyledPostButton>
   );
 };
-
-const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  position: fixed;
-  right: 16px;
-  bottom: 16px;
-  margin-bottom: ${TABBAR_HEIGHT};
-`;
 
 const StyledFAB = styled.button`
   display: flex;
